refactor(api): move get all membership by organization route
in src/team
