+++ api/lib/application/organizations/index.js
@@ -246,41 +246,6 @@ const register = async function (server) {
 
   const orgaRoutes = [
     {
-      method: 'GET',
-      path: '/api/organizations/{id}/memberships',
-      config: {
-        pre: [
-          {
-            method: securityPreHandlers.checkUserBelongsToOrganization,
-            assign: 'belongsToOrganization',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-          query: Joi.object({
-            filter: Joi.object({
-              firstName: Joi.string().empty('').allow(null).optional(),
-              lastName: Joi.string().empty('').allow(null).optional(),
-              email: Joi.string().empty('').allow(null).optional(),
-              organizationRole: Joi.string().empty('').allow(null).optional(),
-            }).default({}),
-            page: Joi.object({
-              number: Joi.number().integer().empty('').allow(null).optional(),
-              size: Joi.number().integer().empty('').allow(null).optional(),
-            }).default({}),
-          }),
-        },
-        handler: organizationController.findPaginatedFilteredMemberships,
-        tags: ['api', 'organizations'],
-        notes: [
-          "Cette route est restreinte aux membres authentifiés d'une organisation",
-          'Elle retourne les rôles des membres rattachés à l’organisation de manière paginée.',
-        ],
-      },
-    },
-    {
       method: 'PATCH',
       path: '/api/organizations/{id}/resend-invitation',
       config: {
