+++ api/src/team/application/membership/membership.route.js
@@ -1,5 +1,6 @@
 import Joi from 'joi';
 
+import { organizationController } from '../../../../lib/application/organizations/organization-controller.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { membershipController } from './membership.controller.js';
@@ -32,7 +33,42 @@ export const membershipRoutes = [
           order: 2,
         },
       },
-      tags: ['api', 'memberships'],
+      tags: ['api', 'team', 'memberships'],
+    },
+  },
+  {
+    method: 'GET',
+    path: '/api/organizations/{id}/memberships',
+    config: {
+      pre: [
+        {
+          method: securityPreHandlers.checkUserBelongsToOrganization,
+          assign: 'belongsToOrganization',
+        },
+      ],
+      validate: {
+        params: Joi.object({
+          id: identifiersType.organizationId,
+        }),
+        query: Joi.object({
+          filter: Joi.object({
+            firstName: Joi.string().empty('').allow(null).optional(),
+            lastName: Joi.string().empty('').allow(null).optional(),
+            email: Joi.string().empty('').allow(null).optional(),
+            organizationRole: Joi.string().empty('').allow(null).optional(),
+          }).default({}),
+          page: Joi.object({
+            number: Joi.number().integer().empty('').allow(null).optional(),
+            size: Joi.number().integer().empty('').allow(null).optional(),
+          }).default({}),
+        }),
+      },
+      handler: organizationController.findPaginatedFilteredMemberships,
+      tags: ['api', 'team', 'organizations'],
+      notes: [
+        "Cette route est restreinte aux membres authentifiés d'une organisation",
+        'Elle retourne les rôles des membres rattachés à l’organisation de manière paginée.',
+      ],
     },
   },
 ];
