+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -523,141 +523,6 @@ describe('Acceptance | Application | organization-controller', function () {
     });
   });
 
-  describe('GET /api/organizations/{id}/memberships', function () {
-    context('Expected output', function () {
-      it('should return the matching membership as JSON API', async function () {
-        // given
-        const adminOfTheOrganization = databaseBuilder.factory.buildUser();
-        const organizationId = databaseBuilder.factory.buildOrganization().id;
-        const adminMembershipId = databaseBuilder.factory.buildMembership({
-          organizationId,
-          userId: adminOfTheOrganization.id,
-          organizationRole: 'ADMIN',
-        }).id;
-
-        const userToUpdateMembership = databaseBuilder.factory.buildUser();
-        const membershipId = databaseBuilder.factory.buildMembership({
-          userId: userToUpdateMembership.id,
-          organizationId: organizationId,
-        }).id;
-
-        await databaseBuilder.commit();
-
-        // when
-        const response = await server.inject({
-          method: 'GET',
-          url: `/api/organizations/${organizationId}/memberships/?filter[email]=&filter[firstName]=&filter[lastName]=&filter[organizationRole]=`,
-          headers: { authorization: generateValidRequestAuthorizationHeader(adminOfTheOrganization.id) },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(200);
-        expect(response.result).to.deep.equal({
-          data: [
-            {
-              attributes: {
-                'organization-role': 'ADMIN',
-              },
-              id: adminMembershipId.toString(),
-              relationships: {
-                user: {
-                  data: {
-                    id: adminOfTheOrganization.id.toString(),
-                    type: 'users',
-                  },
-                },
-              },
-              type: 'memberships',
-            },
-            {
-              attributes: {
-                'organization-role': 'MEMBER',
-              },
-              id: membershipId.toString(),
-              relationships: {
-                user: {
-                  data: {
-                    id: userToUpdateMembership.id.toString(),
-                    type: 'users',
-                  },
-                },
-              },
-              type: 'memberships',
-            },
-          ],
-          included: [
-            {
-              attributes: {
-                email: adminOfTheOrganization.email,
-                'first-name': adminOfTheOrganization.firstName,
-                'last-name': adminOfTheOrganization.lastName,
-              },
-              id: adminOfTheOrganization.id.toString(),
-              type: 'users',
-            },
-            {
-              attributes: {
-                email: userToUpdateMembership.email,
-                'first-name': userToUpdateMembership.firstName,
-                'last-name': userToUpdateMembership.lastName,
-              },
-              id: userToUpdateMembership.id.toString(),
-              type: 'users',
-            },
-          ],
-          meta: {
-            page: 1,
-            pageCount: 1,
-            pageSize: 10,
-            rowCount: 2,
-          },
-        });
-      });
-    });
-
-    context('Resource access management', function () {
-      it('should respond with a 401 - unauthorized access - if user is not authenticated', async function () {
-        // given
-        const organizationId = databaseBuilder.factory.buildOrganization().id;
-        const authorization = 'invalid.access.token';
-
-        // when
-        const response = await server.inject({
-          method: 'GET',
-          url: `/api/organizations/${organizationId}/memberships`,
-          headers: { authorization },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(401);
-      });
-
-      it('should respond with a 403 - forbidden access - if user is not in organization nor is SUPERADMIN', async function () {
-        // given
-        const organizationId = databaseBuilder.factory.buildOrganization().id;
-        const otherOrganizationId = databaseBuilder.factory.buildOrganization().id;
-        const userId = databaseBuilder.factory.buildUser().id;
-        databaseBuilder.factory.buildMembership({
-          organizationRole: Membership.roles.MEMBER,
-          organizationId: otherOrganizationId,
-          userId,
-        });
-
-        await databaseBuilder.commit();
-
-        // when
-        const response = await server.inject({
-          method: 'GET',
-          url: `/api/organizations/${organizationId}/memberships`,
-          headers: { authorization: generateValidRequestAuthorizationHeader(userId) },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(403);
-      });
-    });
-  });
-
   describe('GET /api/admin/organizations/{id}/memberships', function () {
     it('should return the matching membership as JSON API', async function () {
       // given
